--- backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/likes/dto/UserMatchingResponse.java
@@ -1,25 +1,0 @@
-package com.NBE4_5_SukChanHoSu.BE.domain.likes.dto;
-
-import com.NBE4_5_SukChanHoSu.BE.domain.likes.Matching;
-import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.UserProfile;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.Date;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserMatchingResponse {
-    private Long matchingId;
-    private Date matchingTime;
-    private UserProfile user;
-
-    public UserMatchingResponse(UserProfile user, Matching matching) {
-        this.user = user;
-        this.matchingId = matching.getMatchingId();
-        this.matchingTime = matching.getMatchingTime();
-    }
-}
