+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/likes/dto/response/MatchingResponse.java
@@ -1,6 +1,6 @@
 package com.NBE4_5_SukChanHoSu.BE.domain.likes.dto.response;
 
-import com.NBE4_5_SukChanHoSu.BE.domain.likes.Matching;
+import com.NBE4_5_SukChanHoSu.BE.domain.likes.entity.Matching;
 import com.NBE4_5_SukChanHoSu.BE.domain.user.dto.response.UserProfileResponse;
 import lombok.Data;
