+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/likes/dto/response/UserMatchingResponse.java
@@ -1,13 +1,12 @@
 package com.NBE4_5_SukChanHoSu.BE.domain.likes.dto.response;
 
-import com.NBE4_5_SukChanHoSu.BE.domain.likes.Matching;
+import com.NBE4_5_SukChanHoSu.BE.domain.likes.entity.Matching;
 import com.NBE4_5_SukChanHoSu.BE.domain.user.dto.response.UserProfileResponse;
 import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.UserProfile;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.util.Date;
 
 @Data
