+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/likes/entity/Matching.java
@@ -1,4 +1,4 @@
-package com.NBE4_5_SukChanHoSu.BE.domain.likes;
+package com.NBE4_5_SukChanHoSu.BE.domain.likes.entity;
 
 
 import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.UserProfile;
@@ -8,40 +8,37 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import java.util.Date;
+
 
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-@Table( name = "user_likes",
-        indexes = {
-                @Index(name = "idx_like_time", columnList = "like_time")
-        })
-public class UserLikes {
+@Table(name = "matches")
+public class Matching {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userLikeId;
+    private Long matchingId;
 
     @ManyToOne
-    @JoinColumn(name = "liker_id", nullable = false)
+    @JoinColumn(name = "male_user_id", nullable = false)
     @JsonIgnoreProperties({"likes", "likedBy", "maleMatchings", "femaleMatchings","user"})
-    private UserProfile fromUser;  // 좋아요를 보낸 사용자
+    private UserProfile maleUser; // 남자 유저
 
     @ManyToOne
-    @JoinColumn(name = "liked_id", nullable = false)
+    @JoinColumn(name = "female_user_id", nullable = false)
     @JsonIgnoreProperties({"likes", "likedBy", "maleMatchings", "femaleMatchings","user"})
-    private UserProfile toUser;  // 좋아요를 받은 사용자
+    private UserProfile femaleUser; // 두 번째 사용자
 
-    @Column(name = "like_time",nullable = false)
+    @Column(name = "matching_time", nullable = false)
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
-    private Date likeTime;
+    private Date matchingTime; // 매칭된 시간
 
-    public UserLikes(UserProfile fromUser, UserProfile toUser) {
-        this.fromUser = fromUser;
-        this.toUser = toUser;
-        this.likeTime = new Date();
+    public Matching(UserProfile maleUser, UserProfile femaleUser) {
+        this.maleUser = maleUser;
+        this.femaleUser = femaleUser;
+        this.matchingTime = new Date();
     }
 }
