+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/recommend/repository/RecommendUserRepository.java
@@ -1,6 +1,6 @@
-package com.NBE4_5_SukChanHoSu.BE.domain.user.repository;
+package com.NBE4_5_SukChanHoSu.BE.domain.recommend.repository;
 
-import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.RecommendUser;
+import com.NBE4_5_SukChanHoSu.BE.domain.recommend.entity.RecommendUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
