+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/recommend/service/DeleteSchedular.java
@@ -1,6 +1,6 @@
-package com.NBE4_5_SukChanHoSu.BE.domain.user.service;
+package com.NBE4_5_SukChanHoSu.BE.domain.recommend.service;
 
-import com.NBE4_5_SukChanHoSu.BE.domain.user.repository.RecommendUserRepository;
+import com.NBE4_5_SukChanHoSu.BE.domain.recommend.repository.RecommendUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
-public class RecommendService {
+public class DeleteSchedular {
     private final RecommendUserRepository recommendUserRepository;
 
     @Scheduled(cron = "0 0 0 * * ?")    // 매일 자정
